modularized statefulValidation
